test: align session export checks with rewrap primitive
